dailies: prevent matching .org_archive files
`org-roam-dailies--list-files` would return `.org_archive` files along with `.org` files. The real issue is that `org-roam-dailies-goto-previous-note` can step into the archive file but not step out of the archive file. This could be fixed by not visiting the archive file or by fixing whatever issue prevents `org-roam-dailies-goto-previous-note` from stepping out of the archive file. this commit updates it to not consider archive files.