perf(db): Just check mtime instead of sha1 hashes
To be clear, the actual workhorse `org-roam-db-update-file` still
checks and records SHA1 hashes.

This change just applies to when `org-roam-db-sync` determines whether
any files have changed, in order to run `org-roam-db-update-file` on
them.

If the mtime has changed but the content has not, then
`org-roam-db-update-file` will still refuse to re-scan it.